Fix diff list
The diff list was being generated from the git tag list when any
reference link was found. However there are two problems with this:

 - The tag list could be generated in the middle of the markdown if a
   reference link was found before the end of the text
 - The tag reference was formatted was a tag instead of a version. For
   example, instead of showing `[1.0.0]:` it would show `[v1.0.0]`

The diff list is now being generated from the versions found in the
changelog. Every reference link is deleted and the diff list is
generated at the end of the file.

Some reference links for other purposes could be deleted with this
design but for now it's a good enough compromise.