Change to accept unknown server
When an unknown server was detected the release command would not
complete but there is no logical reason for that to happen as a change
log does not have to have a diff list. Instead if a server is unknown
the release command simply does not write the diff list.