Split repository and connection URLs
The connection URL was being used to generate diff links but doing so
was flawed because the web page URL and Git URL can be different. Also
the connection URL can have GIT/SSH protocol while such protocols don't
make sense in a web page.

With this modification having an SCM connection URL with GIT or SSH
protocol does not result in an error.