Change set difference method names
Methods that performed a set difference had a `Without` infix . However
`SetAWithoutSetB` is kind of hard to understand as a set difference.
The most common infix for a set difference operation is `Except` since
that's what SQL standard called it. In Java a set difference is called
`removeAll` but that name may falsely presume that these are a mutable
operations when these operations are actually immutable. Not in a proper
way (it's implemented as a copy and remove all) but still immutable.